restructure forc-crypto main file, introduce lib file
